+++ src/main/java/org/jabref/preferences/PreferencesService.java
@@ -1,6 +1,5 @@
 package org.jabref.preferences;
 
-import java.nio.charset.Charset;
 import java.nio.file.Path;
 import java.util.List;
 import java.util.Map;
@@ -68,8 +67,6 @@ public interface PreferencesService {
 
     FilePreferences getFilePreferences();
 
-    void storeFilePreferences(FilePreferences filePreferences);
-
     FieldWriterPreferences getFieldWriterPreferences();
 
     FileHistory getFileHistory();
@@ -97,8 +94,6 @@ public interface PreferencesService {
     SavePreferences getSavePreferencesForExport();
 
     SavePreferences getSavePreferences();
-
-    Charset getDefaultEncoding();
 
     String getUser();
 
@@ -138,8 +133,6 @@ public interface PreferencesService {
     Language getLanguage();
 
     void setLanguage(Language language);
-
-    BibDatabaseMode getDefaultBibDatabaseMode();
 
     GeneralPreferences getGeneralPreferences();
 
@@ -183,11 +176,7 @@ public interface PreferencesService {
 
     RemotePreferences getRemotePreferences();
 
-    void storeRemotePreferences(RemotePreferences remotePreferences);
-
     ProxyPreferences getProxyPreferences();
-
-    void storeProxyPreferences(ProxyPreferences proxyPreferences);
 
     //*************************************************************************************************************
     // CitationKeyPatternPreferences
@@ -305,12 +294,17 @@ public interface PreferencesService {
 
     GuiPreferences getGuiPreferences();
 
-    void storeGuiPreferences(GuiPreferences guiPreferences);
-
     void clearEditedFiles();
 
+    /**
+     * Gets the directory for file browsing dialogs. This ensures that each browse dialog starts in the last visited
+     * browse directory.
+     */
     Path getWorkingDir();
 
+    /**
+     * Stores the directory for file browsing dialogs
+     */
     void setWorkingDirectory(Path dir);
 
     //*************************************************************************************************************